# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Platform API
  description: |
    社内ナレッジ活用プラットフォーム — BFF + マイクロサービス統合仕様書
    FR-01, FR-03, FR-04, FR-05, FR-06, FR-07, FR-08, FR-09, FR-10, FR-11, FR-13, FR-22,
    UC-01, UC-02, UC-05, UC-11
  version: "0.1.0"
  contact:
    name: Platform Team
servers:
  - url: http://localhost:5000
    description: BFF (ローカル開発)
  - url: http://localhost:5001
    description: DocumentService (直接)
  - url: http://localhost:5003
    description: RetrievalService (直接)
  - url: http://localhost:5004
    description: AiAnalysisService (直接)
  - url: http://localhost:5005
    description: AuthorizationService (直接)
  - url: http://localhost:5007
    description: "LlmGateway — IADR-0017: 内部 API。通常は host 非公開でサービス間呼び出し専用（compose では expose のみ）。この URL は個別に dotnet run した場合のみ到達可能"
...
paths:
...
  /dashboard/events:
    post:
      tags: [Dashboard]
      summary: "FR-10: 利用イベント（検索・回答）を記録する"
      description: |
        集計の入力となるため、認証済みなら誰でも記録できる（認証は必須）。
        利用者は JWT から特定する（テスト・開発環境では anonymous）。
      operationId: record-usage-event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageEventRequest"
      responses:
        "201":
          description: 記録された利用イベント
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageEventCreatedDto"
        "400":
          description: "eventType 不正（search / answer 以外）"
        "401":
          description: 未認証
...
  /bff/graph/edge-types:
    get:
      # NFR-09: 認証のみ・ロール不問。
      # 🔴 **後段は /graph/edge-types/catalog（認証のみ・使用件数なし）であって、
      # /graph/edge-types（admin / operator 限定・使用件数つき）ではない。**
      # 後者へ向けると一般利用者が 403 になり、かつ ABAC で絞られていない集計値が漏れる。
      x-roles: []
      tags: [Graph]
      summary: "FR-17, SC-18: 辺の型カタログ（描画用）"
      description: |
        FR-17, SC-18, ADR-0033, ADR-0039: 辺の型の識別子・名前・層・対称性を返す。

        グラフ読み取りの応答が返すのは `edgeTypeId` だけなので、**辺の描き分けと型フィルタには
        このカタログが要る**（表示名は辞書側で解決する。改名に追随させるため。ADR-0033 決定 9）。

        **使用件数は含まない。** 件数は全辺を数えており ABAC で絞られていないため、
        一般利用者へ返すと権限外の辺を含む総量が漏れる。件数が要る管理用途（SC-09 / SC-10）は
        admin / operator 限定の別口が担う。
      operationId: bff-graph-edge-types
      responses:
        "200":
          description: 辺の型カタログ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EdgeTypeCatalogItem" }
        "403": { description: 後段が拒否した（そのまま透過する） }
  /bff/graph/{documentId}:
    parameters:
      - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      # NFR-09: 認証のみ・ロール不問。**可視性はロールではなく ABAC が決める** ——
      # GraphService が JWT から解決したスコープでホップごとに判定する（ADR-0034 決定 1）。
      x-roles: []
      tags: [Graph]
      summary: "FR-17, UC-10: 起点ノード 1 件（グラフ読み取り）"
      description: |
        FR-17, UC-10, ADR-0034: 起点となる文書のノードを 1 件返す。

        **権限外・複製未到達・不存在はすべて 404 に倒す**（ADR-0034 決定 2）。
        利用者は「リンク切れ」と「権限不足」を区別できない —— これは受け入れ済みの副作用である。
        403 は返さない（403 は「権限が無いだけで存在はする」ことを漏らす）。
      operationId: bff-graph-node
      responses:
        "200":
          description: グラフ（ノード 1 件）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphView" }
        "404": { description: 権限外・不存在（区別しない） }
  /bff/graph/{documentId}/neighbors:
    parameters:
      - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      # クエリ引数は **操作（get）の下**へ置く。path レベルへ置くと orval が
      # 空の分解代入を含む URL 組み立てを生成し、noUnusedLocals（TS6198）で型検査が落ちる（実測）。
      parameters:
        - name: hops
          in: query
          required: false
          description: "探索の深さ。既定 2 / 上限 3（ADR-0034 決定 3）。上限超過は 400 で拒否し、黙って切り詰めない"
          schema: { type: integer, minimum: 1, maximum: 3 }
        - name: by
          in: query
          required: false
          description: |
            間引きの基準（ADR-0049 決定 4）。表示上限（ノード 200）を超えた候補集合から何を残すか。
            **未知の値・未指定は既定（distance）へ縮退する** —— 例外にしない。
          schema: { type: string, enum: [distance, updated, degree] }
        - name: types
          in: query
          required: false
          description: |
            SC-18 (#917): 辺の型フィルタ（型 ID のカンマ区切り。未指定・空 = 絞らない）。
            🔴 **絞りはサーバ側で探索の入口に適用される**（planning#446）——クライアントで
            打ち切り後に絞ると「上位 200 件のうち一致したもの」になり範囲が意図せず狭まる。
            総数（totalNodes / totalEdges）もフィルタ後の母集合で数え直される。
            GUID として読めない要素は 400（`edge_type_filter_invalid`）。**実在しない型 ID は
            エラーにならず、単に 1 本も一致しない**（型辞書は公開済みの語彙であり秘匿対象ではない）。
          schema: { type: string }
      # NFR-09: 認証のみ・ロール不問（上と同じ。可視性は ABAC が決める）。
      x-roles: []
      tags: [Graph]
      summary: "FR-17, UC-10: 近傍探索（ホップごとに権限を判定する）"
      description: |
        FR-17, UC-10, ADR-0034: 起点から関係を辿った近傍グラフを返す。

        **判定はホップごとに行い、権限外ノードは展開の前に刈る**（決定 1）。
        したがって権限外ノードが「橋」となって権限外の隣接文書が浮上することはない。
        **表示上限（ノード 200 / 辺 500）の計数は権限判定を通過した品目に対してのみ行う**ため、
        件数から権限外文書の存在が漏れることもない（決定 4）。
      operationId: bff-graph-neighbors
      responses:
        "200":
          description: 近傍グラフ
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphView" }
        "400": { description: hops が範囲外（1〜3） }
        "404": { description: 権限外・不存在（区別しない） }
  /bff/graph/suggestions:
    get:
      parameters:
        - name: state
          in: query
          required: false
          description: |
            SC-21 (#918): 状態フィルタ。**未指定は `pending`（既定）**であり、`all` は絞りを外す。
            🔴 **`all` は状態の値ではない** —— 永続層に `all` という状態は存在しない。
            未知の値は 400（`invalid_state`）。**既定の補完も値域の検査も後段が一箇所で持つ**
            （BFF は透過するだけである）。
          schema: { type: string, enum: [pending, approved, rejected, all] }
        - name: kind
          in: query
          required: false
          description: |
            SC-21 (#918): 種類フィルタ。未指定は絞らない（＝すべて）。
            🔴 **リンク提案とタグ提案は同一の一覧に同居させる。画面を分けない**（FR-18）。
            未知の値は 400（`invalid_kind`）。
          schema: { type: string, enum: [link, tag] }
      # NFR-09: 認証のみ・ロール不問。**可視性はロールではなく ABAC が決める** ——
      # 権限のない文書に関する提案は一覧にも件数にも現れない（05_screens §SC-21 アクセス制御）。
      x-roles: []
      tags: [Graph]
      summary: "FR-18, UC-10, SC-21: AI 提案の一覧（棚卸し用・読み取りのみ）"
      description: |
        FR-18, UC-10, SC-21, ADR-0033 決定 7・10: AI が提案したリンク候補・タグ候補を一覧で返す。

        🔴 **読み取りだけである。承認・却下の口は BFF に無い。**
        承認の主導線は SC-03（文書詳細）であり、SC-21 は棚卸し用の**従**である。
        **一括承認の手段は画面・API のいずれにも設けない**（FR-18。一覧の 1 行に収まる情報では
        承認を判断できず、タイトルだけを見て機械的に承認する運用に落ちるためである）。

        **権限のない文書に関する提案は、件数を含め一切現れない**（存在秘匿）。
        後段が引けないときは 502 であり、**空配列へ縮退しない** ——
        「提案が 1 件も無い」と「一覧が引けない」は利用者にとって別の意味である。
      operationId: bff-graph-suggestions
      responses:
        "200":
          description: AI 提案の一覧（権限内のもののみ）
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AiSuggestion" }
        "400": { description: state / kind が値域外（そのまま透過する） }
components:
  schemas:
...
    GraphNodeItem:
      type: object
      description: |
        FR-17: グラフのノード（文書単位）。
        `isPrivateNote` は SC-18 の描き分け（組織文書＝円＋📄 / 個人資料＝角丸四角＋👤）のための
        1 bit（ADR-0054 / #917）。値が付くのは ABAC 判定を通過した可視ノードだけであり、
        **doc_scope を持たない文書は組織文書（false）として返る**（ADR-0054 決定 5）。
      required: [documentId, title]
      properties:
        documentId: { type: string, format: uuid }
        title: { type: string }
        isPrivateNote: { type: boolean, default: false }
    GraphEdgeItem:
      type: object
      description: |
        FR-17, ADR-0033 決定 4: グラフの辺。**辺の型は識別子で返す**（表示名は辞書側で解決する。
        改名に追随するため）。`provenance` は出所（auto / user / ai-approved）。
      required: [id, sourceDocumentId, targetDocumentId, edgeTypeId, provenance]
      properties:
        id: { type: string, format: uuid }
        sourceDocumentId: { type: string, format: uuid }
        targetDocumentId: { type: string, format: uuid }
        edgeTypeId: { type: string, format: uuid }
        provenance: { type: string, enum: [auto, user, ai-approved] }
    EdgeTypeCatalogItem:
      type: object
      description: |
        FR-17, SC-18: 辺の型（描画用）。**使用件数を持たない** —— 意図した除外である
        （件数は ABAC で絞られておらず、一般利用者へ返すと総量が漏れる）。
      required: [id, name, layer, isSymmetric]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, description: "表示名。改名されても id は変わらない" }
        layer: { type: string, description: "型の層（core / domain / user 等）" }
        isSymmetric: { type: boolean, description: "対称な関係か（向きを持たないか）" }
    AiSuggestion:
      type: object
      description: |
        FR-18, SC-21, ADR-0033 決定 7・10: AI 提案（リンク候補・タグ候補）。

        **リンク提案とタグ提案を 1 つの型で表す** —— SC-21 が同一の一覧を求めており、
        型を分けるとクライアント側で 2 経路に割れる。種別固有の欄は省略可である。

        🔴 **本文指紋（却下解除の判定に使う内部状態）は公開面に出さない。**

        🔴 **辺の型の表示名を持たない。** 名前は辞書（`/bff/graph/edge-types`）で解決し、
        改名に追随させる（ADR-0033 決定 9）。ここへ焼き込むと改名後も古い名前を出し続ける。
      required:
        [id, kind, sourceDocumentId, rationale, state, rejectedCount, sourceDocumentTitle]
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          description: "link（リンク候補）/ tag（タグ候補）"
        sourceDocumentId:
          type: string
          format: uuid
          description: "リンク提案では起点、タグ提案では対象文書。**どちらの種別でも必ず入る**"
        targetDocumentId:
          type: string
          format: uuid
          nullable: true
          description: "リンク提案のみ"
        edgeTypeId:
          type: string
          format: uuid
          nullable: true
          description: "リンク提案のみ。表示名は辺の型カタログで解決する"
        tagValue:
          type: string
          nullable: true
          description: "タグ提案のみ。SC-09 のタグ辞書に整合する値であること"
        rationale: { type: string, description: "なぜ関連と判断したか（SC-21 主要素 6）" }
        state:
          type: string
          description: "pending（承認待ち）/ approved（承認済み）/ rejected（却下）"
        rejectedCount: { type: integer, description: "累積の却下回数" }
        reinstatedReason:
          type: string
          nullable: true
          description: >-
            再提示の理由。source / target / both のいずれか。**非 null なら再提示であり、
            画面は固定文言を必ず添える**（ADR-0033 決定 10。理由の無い再提示を起こさない）
        sourceDocumentTitle:
          type: string
          description: "起点（タグ提案では対象）の文書名。SC-21 の「提案の内容」列が要る"
        targetDocumentTitle:
          type: string
          nullable: true
          description: "終点の文書名。タグ提案では終点が無いため null である"
    GraphView:
      type: object
      description: |
        FR-17, UC-10, ADR-0034 決定 4: 近傍グラフ。`truncated` は表示上限で打ち切ったかを表す。
        **上限の計数は権限判定を通過した品目に対してのみ行われる**ため、この値が権限外文書の
        存在を漏らすことはない。
      required: [nodes, edges, truncated]
      properties:
        nodes:
          type: array
          items: { $ref: "#/components/schemas/GraphNodeItem" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/GraphEdgeItem" }
        totalNodes:
          type: integer
          description: |
            FR-17, ADR-0049 決定 1・3: **許可済み**ノードの総数。表示上限（200）を超えて数えた値。
            🔴 **権限外は 1 件も含まない** —— 数えてよいのは ABAC 判定を通過した品目だけである。
        totalEdges:
          type: integer
          description: "許可済みの辺の総数（同上）"
        totalIsLowerBound:
          type: boolean
          description: |
            ADR-0049 決定 3: 算出用の上限（ノード 2,000 / 辺 5,000）に達したか。
            **立っていれば総数は「以上」の意味である**（例:「2,000 件以上」）。
            ⚠️ このとき `updated` / `degree` の並びは**厳密な上位 200 件ではない**（同 決定 4）。
        truncated: { type: boolean }
...
    UsageEventRequest:
      type: object
      required: [eventType]
      properties:
        eventType:
          type: string
          enum: [search, answer]
          description: "検索実行（search）/ AI 回答生成（answer）"
        query:
          type: string
          nullable: true
          description: "検索語（種別が search のとき任意。検索傾向の集計に用いる）"
    UsageEventCreatedDto:
      type: object
      description: |
        FR-10: 記録された利用イベントの ID。契約 record を持たず、`DashboardEndpoints` の匿名型
        （`new { ev.Id }`）がそのまま JSON になる。`UsageEvent.Id` は非 null の `Guid`。
      required: [id]
      properties:
        id: { type: string, format: uuid }
...